requirements.txt
pytest==7.4.0
PyYAML==6.0
torch==2.0.1
